% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{GreetingsFrom}
\alias{GreetingsFrom}
\title{Greetings from who}
\usage{
GreetingsFrom(who)
}
\arguments{
\item{who}{"Tek"}
}
\value{
who says hello
}
\description{
Greetings from Tek
}
\examples{
GreetingsFrom("Tek")
}
