% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{welcome}
\alias{welcome}
\title{Greetings from Tek}
\usage{
welcome(who)
}
\arguments{
\item{who}{"Tek"}
}
\value{
who says hello
}
\description{
Greetings from Tek
}
\examples{
bi_to_dec("Tek")
}
